Port settings to new nixpkgs version
Moved plasma setting, since it was changed. Renamed another zsh config
variable. Set a warning related to nix versions to false. Possible this
is going to shoot myself in the foot later, but its really annoying that
home-manager and nixpkgs can both be on unstable and still complain they
aren't synched. They are expected to work fine in this way.